chatbot/requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.3.1
executing==2.2.1
frozenlist==1.8.0
genai==2.1.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-genai==1.64.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==8.38.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
multidict==6.7.1
narwhals==2.16.0
numpy==2.2.6
packaging==26.0
pandas==2.3.3
parso==0.8.6
pillow==12.1.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.54.0
tabulate==0.9.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.6.0
websockets==15.0.1
yarl==1.22.0
